refactor: create multihash_impl
The `Multihash` proc macro no longer creates a code table, hence the
name `code` doesn't really apply anymore. Thus is it was renamed to
`multihash_impl.rs` which is kind of symmetric to `hasher_impl.rs`.

The proc macro is moved to a `derive` module, to have a cleaner
distinction between importing the default `Multihash` implementation
or the `Multihash` derive.
