+++ src/multihash_impl.rs
@@ -44,58 +44,58 @@ pub const STROBE_512: u64 = 0xa1;
 #[derive(Clone, Debug, Eq, Multihash, PartialEq)]
 pub enum Multihash {
     /// Multihash array for hash function.
-    #[mh(code = crate::IDENTITY, hasher = crate::Identity256)]
+    #[mh(code = self::IDENTITY, hasher = crate::Identity256)]
     Identity256(crate::IdentityDigest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA1, hasher = crate::Sha1)]
+    #[mh(code = self::SHA1, hasher = crate::Sha1)]
     Sha1(crate::Sha1Digest<crate::U20>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA2_256, hasher = crate::Sha2_256)]
+    #[mh(code = self::SHA2_256, hasher = crate::Sha2_256)]
     Sha2_256(crate::Sha2Digest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA2_512, hasher = crate::Sha2_512)]
+    #[mh(code = self::SHA2_512, hasher = crate::Sha2_512)]
     Sha2_512(crate::Sha2Digest<crate::U64>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA3_224, hasher = crate::Sha3_224)]
+    #[mh(code = self::SHA3_224, hasher = crate::Sha3_224)]
     Sha3_224(crate::Sha3Digest<crate::U28>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA3_256, hasher = crate::Sha3_256)]
+    #[mh(code = self::SHA3_256, hasher = crate::Sha3_256)]
     Sha3_256(crate::Sha3Digest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA3_384, hasher = crate::Sha3_384)]
+    #[mh(code = self::SHA3_384, hasher = crate::Sha3_384)]
     Sha3_384(crate::Sha3Digest<crate::U48>),
     /// Multihash array for hash function.
-    #[mh(code = crate::SHA3_512, hasher = crate::Sha3_512)]
+    #[mh(code = self::SHA3_512, hasher = crate::Sha3_512)]
     Sha3_512(crate::Sha3Digest<crate::U64>),
     /// Multihash array for hash function.
-    #[mh(code = crate::KECCAK_224, hasher = crate::Keccak224)]
+    #[mh(code = self::KECCAK_224, hasher = crate::Keccak224)]
     Keccak224(crate::KeccakDigest<crate::U28>),
     /// Multihash array for hash function.
-    #[mh(code = crate::KECCAK_256, hasher = crate::Keccak256)]
+    #[mh(code = self::KECCAK_256, hasher = crate::Keccak256)]
     Keccak256(crate::KeccakDigest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::KECCAK_384, hasher = crate::Keccak384)]
+    #[mh(code = self::KECCAK_384, hasher = crate::Keccak384)]
     Keccak384(crate::KeccakDigest<crate::U48>),
     /// Multihash array for hash function.
-    #[mh(code = crate::KECCAK_512, hasher = crate::Keccak512)]
+    #[mh(code = self::KECCAK_512, hasher = crate::Keccak512)]
     Keccak512(crate::KeccakDigest<crate::U64>),
     /// Multihash array for hash function.
-    #[mh(code = crate::BLAKE2B_256, hasher = crate::Blake2b256)]
+    #[mh(code = self::BLAKE2B_256, hasher = crate::Blake2b256)]
     Blake2b256(crate::Blake2bDigest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::BLAKE2B_512, hasher = crate::Blake2b512)]
+    #[mh(code = self::BLAKE2B_512, hasher = crate::Blake2b512)]
     Blake2b512(crate::Blake2bDigest<crate::U64>),
     /// Multihash array for hash function.
-    #[mh(code = crate::BLAKE2S_128, hasher = crate::Blake2s128)]
+    #[mh(code = self::BLAKE2S_128, hasher = crate::Blake2s128)]
     Blake2s128(crate::Blake2sDigest<crate::U16>),
     /// Multihash array for hash function.
-    #[mh(code = crate::BLAKE2S_256, hasher = crate::Blake2s256)]
+    #[mh(code = self::BLAKE2S_256, hasher = crate::Blake2s256)]
     Blake2s256(crate::Blake2sDigest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::STROBE_256, hasher = crate::Strobe256)]
+    #[mh(code = self::STROBE_256, hasher = crate::Strobe256)]
     Strobe256(crate::StrobeDigest<crate::U32>),
     /// Multihash array for hash function.
-    #[mh(code = crate::STROBE_512, hasher = crate::Strobe512)]
+    #[mh(code = self::STROBE_512, hasher = crate::Strobe512)]
     Strobe512(crate::StrobeDigest<crate::U64>),
 }
